Refactor and clean-up
BoardState to be used to manage all board data in the game. Previous methods removed/renamed and replaced with more clear, comprehensive methods. "update" property added and initialized to false to toggle between methods. "cellCombinations" object moved from prior BoardAssembly class to new BoardState class as a property to manage the state of the cells. getRows() method removed. "line" property moved from the BoardPersistence class to be used in the BoardPrinter() class to manage the formatting of the board. All commented code in process of being majorly refactored. Current implementations pass in test suite.