Refactor board.js implementations and tests
Board class refactored to only manage the state of the Tic-Tac-Toe board, accept a given number of columns, have a 'cells' property that initializes a new nested array of rows and columns, and provide information on the total number of columns passed to the class. getRows() method removed. printBoard() method removed. Logic of these two methods now persists in newly created Display class. Display class responsible for managing the format/construction of the board based on the instance of the board state object passed to it. The line  property of pre-existing Board class moved to Display class, refactored to use the repeat() method to create a dashed line, and renamed as the rowSeparator.

Board tests refactored to fit the refactored implementations of the Board Class and Display Class. All tests passing and in working state. Refactored board assembly method in progress.